+++ packages/node-sdk/src/index.ts
@@ -1,3 +1,6 @@
 export * from './types.js';
 export * from './sdk.js';
 export * from './crypto/core.js';
+export * from './middleware/common.js';
+export * from './middleware/express.js';
+export * from './middleware/fastify.js';